+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @weaver-conf/playground
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @weaver-conf/weaver-server@0.1.2
+  - @weaver-conf/config-types@0.1.1
+  - @weaver-conf/storage-provider-static-json@0.1.1
+  - @weaver-conf/weaver-client@0.1.1
+
 ## 0.1.1
 
 ### Patch Changes